requirements.txt
groq==0.4.1
google-generativeai==0.3.0
python-binance==1.0.17
python-telegram-bot==20.1
APScheduler==3.10.4
python-dotenv==1.0.0
Flask==3.0.0
